EDIT: all components refactoring
