+++ src/components/Footer/Footer.js
@@ -1,0 +1,12 @@
+import './footer.scss';
+
+const Footer = () => {
+  return (
+    <footer className="footer">
+      <p className="footerInformation">위스타그램 정보입니다.</p>
+      <p className="footerMakeYear">2022년도에 제작</p>
+    </footer>
+  );
+};
+
+export default Footer;
