+++ src/components/Footer/footer.scss
@@ -1,0 +1,13 @@
+.footer {
+  grid-area: footer;
+
+  .footerInformation {
+    opacity: 0.3;
+    font-weight: 700;
+    font-size: 1.3rem;
+  }
+
+  .footerMakeYear {
+    margin-top: 1rem;
+  }
+}
