+++ src/components/Header/Header.js
@@ -1,3 +1,5 @@
+import './header.scss';
+
 const Header = () => {
   return (
     <header className="topArea">
