+++ src/components/Header/header.scss
@@ -1,0 +1,53 @@
+.topArea {
+  grid-area: top;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  position: fixed;
+  width: 100%;
+  height: 10vh;
+  padding: 1rem 0;
+  background-color: #ffffff;
+
+  .logo {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 20%;
+
+    img {
+      width: 3rem;
+      height: 3rem;
+    }
+
+    span {
+      display: flex;
+      justify-content: center;
+      width: 10rem;
+      margin-left: 1rem;
+      border-left: 0.2rem solid;
+      text-decoration: none;
+      font-size: 2rem;
+      font-family: 'Lobster', cursive;
+    }
+  }
+
+  [type='search'] {
+    width: 20%;
+    height: 70%;
+    text-align: center;
+    opacity: 0.5;
+  }
+
+  .icon {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    width: 20%;
+
+    img {
+      width: 3rem;
+      height: 3rem;
+    }
+  }
+}
