+++ src/pages/minseok/Main/components/CommentList/commentList.scss
@@ -1,0 +1,16 @@
+.postingCommentList {
+  .userCommentId {
+    font-weight: 700;
+    font-size: 1.6rem;
+  }
+
+  .userComment {
+    margin-top: 0.5rem;
+    margin-left: 1rem;
+    font-size: 1.3rem;
+  }
+
+  .userCommentRemove {
+    margin-left: 1rem;
+  }
+}
