+++ src/pages/minseok/Main/components/Fallow/Fallow.js
@@ -1,3 +1,5 @@
+import './fallow.scss';
+
 const Fallow = () => {
   return (
     <div className="sidebarFallow">
