+++ src/pages/minseok/Main/components/History/history.scss
@@ -1,0 +1,61 @@
+.sidebarHistory {
+  margin-bottom: 1rem;
+  padding: 1rem;
+  border: 0.1rem solid #e9e9e9;
+  background-color: #ffffff;
+
+  .sidebarHistoryTitle {
+    display: flex;
+    margin-bottom: 1rem;
+
+    h1 {
+      opacity: 0.3;
+      font-size: 1.6rem;
+      font-weight: 700;
+    }
+
+    button {
+      margin-left: auto;
+      font-size: 1.3rem;
+      font-weight: 700;
+    }
+  }
+
+  .sidebarHistoryUser {
+    display: flex;
+    flex-direction: column;
+
+    li:last-child {
+      margin-bottom: 0;
+    }
+
+    .sidebarHistoryUserList {
+      display: flex;
+      align-items: center;
+      margin-bottom: 1rem;
+
+      .sidebarHistoryUserListImage {
+        width: 3rem;
+        height: 3rem;
+        padding: 0.1rem;
+        border: 0.2rem solid #ff0000;
+        border-radius: 50%;
+      }
+      .sidebarHistoryUserListId {
+        margin-left: 1rem;
+
+        .sidebarHistoryUserListIdName {
+          font-size: 1.3rem;
+          font-weight: 700;
+        }
+
+        .sidebarHistoryUserListIdTime {
+          margin-top: 0.5rem;
+          opacity: 0.3;
+          font-size: 1.3rem;
+          font-weight: 700;
+        }
+      }
+    }
+  }
+}
